Move log out of the helper package 🌿
