+++ pkg/cmd/pipelinerun/logs.go
@@ -21,10 +21,10 @@ import (
 
 	"github.com/spf13/cobra"
 	"github.com/tektoncd/cli/pkg/cli"
-	"github.com/tektoncd/cli/pkg/helper/log"
 	"github.com/tektoncd/cli/pkg/helper/options"
 	prhelper "github.com/tektoncd/cli/pkg/helper/pipelinerun"
 	validate "github.com/tektoncd/cli/pkg/helper/validate"
+	"github.com/tektoncd/cli/pkg/log"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
